README.md
The academic performance of students is a crucial aspect of education. This study seeks to examine the  factors: parental level of education, lunch and test preparation effect on academic performance. The study  used a quantitative method for its research design. A random sampling technique was used of which the sample consisted of 100 students. Regression analysis was used to identify the effect of independent variables (factors) on students’ performance using the R studio software.